[dev.typeparams] cmd/compile: don't rename closures that have a name
If a closure has already been given a name, don't assign it a new one.

The general handling of function literals and naming between all the
different legacy IR construction code paths is a mess. E.g., the exact
semantics around who's responsible for assigning the underlying
function's linker name or adding it to typecheck.Target.Decls.

Probably we should try to simplify and unify this logic too. E.g.,
enforce an invariant that the underlying function is included in
typecheck.Target.Decls iff it has a non-blank name.

Change-Id: I7da1e93c70d268f575b12d6aaeb2336eb910a6f1