[dev.typeparams] cmd/compile: add -d=unified flag to enable unified IR
This CL adds a new -d=unified debug flag, which controls whether
unified IR mode is used. 0 is off, 1 is on, and 2 is on with generics
support. ("-G -d=unified" is also supported for generics + unified IR,
but test/run.go doesn't support passing multiple compiler flags to "go
build" or "go run".

Change-Id: Iaa5f3cc0a24b9881aeec5317cd6b462b4a7b6fc9