[dev.typeparams] cmd/compile: load code pointer during call expansion
DO NOT SUBMIT

This CL changes OpClosureLECall to only have a closure argument, but
instead having an Offset field in the AuxCall to distinguish what
offset in the closure argument the pointer should be loaded. E.g., 0
for a normal function closure, or a per-method index for interface
calls.

Ops for loading the Offset field are generated during late call
expansion. This evidently helps to minimize the register pressure, as
cmd/compile's overhead shrinks from 4kB to ~400 bytes.

The next optimization seems to be changing AMD64 so that instead of
generating:

	MOVQ itab(SB), DX
	MOVQ 24(DX), DI
	CALL DI

we generate:

	MOVQ itab(SB), DX
	CALL 24(DX)

This appears to work with cmd/asm, but I don't know enough Go assembly
to tell if this is supported by other CPUs too.

Change-Id: Idbc8855d0e7c8c3038092371fbb9ba3ce447dcca