cmd/cgo: output cgo pragmas as object instead of go
In this commit, cmd/cgo now outputs the -dynimport file into an object
file if the -dynobjout flag is used instead of the -dynout flag.

cmd/go is updated accordingly to directly use the object file instead of
using the written Go file.

This pull request updates issue #15681. It should allow the cgo
compiling processes to be done differently so that it can be faster. It
should also make it easier for gcc to be parallelized.

cc @mdempsky

Change-Id: I1f73bba475cdc78a73cdc40bd02f8b70a297a66b
GitHub-Last-Rev: 890ad67fba53661f44c0a9136006390cc18bcb31
GitHub-Pull-Request: golang/go#46797