cmd/compile/internal/noder: fix a deadlock when compiling large packages of incorrect files
When compiling large packages of files that are syntactically correct, but otherwise incorrect (types, etc...), a deadlock happens because the line 62 `syntax.Parse` is trying to push multiple errors to the unbuffered error channel on the node, stopping the goroutines, and halting the main routine at line 68 waiting to range over an open channel.

Buffering the channel to allow for multiple errors to be pushed to the node fixes this problem.

Change-Id: I536bf4ab50e2aee29237ac1bad862c05c57d9162
GitHub-Last-Rev: fcc04cf15b33e6774a097d02c059932229e2d690
GitHub-Pull-Request: golang/go#46808