cmd/asm, cmd/internal/obj: add initial support for loong64 assembler
This is preparatory work towards a full-fledged loong64 port, plus the
initial import of LoongArch instruction data and encoding tests for
canonical forms of almost all supported instructions (except jumps,
which are to be implemented later because of the necessary scaffolding).

Instruction encodings and corresponding tests are generated with data
from the https://github.com/loongson-community/loongarch-opcodes repo,
at commit 0850ee22fd62.

The end-to-end test encodings are manually checked against objdump
output. The binutils port from Loongson is definitely WIP and not
upstream yet, but is already open-sourced at
https://github.com/loongson/binutils-gdb. This port is not incorporating
bits from the loongarch-opcodes repo yet, so beware that mnemonics and
syntax can differ.

Updates #46229

Change-Id: Iac90c3e81feb3f7190a3b4583692b63c8c4bf35d