cmd/go: call modload.WriteGoMod explicitly, once in each command
Previously, WriteGoMod was called implicitly in several modload
functions. For some commands, it would be called multiple times which
might leave it in a bad intermediate state if an error occurred later
in the command. {Allow,Disallow}WriteGoMod could be used to determine
whether it could be written.

This change makes writing go.mod and go.sum more
straightforward. Commands should explicitly call modload.WriteGoMod
once when modules are enabled. WriteGoMod now returns an error instead
of calling base.Fatalf. WriteGoMod returns an error if changes are
needed but the command is in readonly or vendor mode. WriteGoMod may
be called safely outside a module.

For #45551

Change-Id: I07ef35450ac8e75b0f2d89e05978aa699fcf3be7