[dev.typeparams] cmd/compile: start using sub-dictionary entries where needed
Added new struct instInfo for information about an instantiation (of a
generic function/method with gcshapes or concrete types). We use this to
remember the dictionary param node, the nodes where sub-dictionaries
need to be used, etc.

Added code to start accessing sub-dictionary entries at the appropriate
call sites. These compile, but don't work yet, because we're not filling
in actually sub-dictionary entries yet.

Other improvements:
 - Only create sub-dictionary entries when the function/method
   call/value or closure actually has type params in it. Added new test
   file subdict.go with an example where a generic method has an
   instantiated method call that does not depend not have type params.

 - Added in code to deal with "generic" closures. This includes
   calculating instInfo for an instantiated closure. For simplicity, we
   are giving the closure the same type params as the containing generic
   function. But it's still missing some code to capture the needed
   dictionary from the containing outer function.

Change-Id: I691b9dc024a89d2305fcf1d8ba8540e53c9d103f