[dev.typeparams] cmd/compile: basic dictionary-passing prototype
DO NOT REVIEW
DO NOT SUBMIT

This is a very minimal demonstration of using "hidden parameters" to
pass an extra parameter to functions and methods. For this CL, the
hidden argument is just always initialized to zero.

Uploading just because it's working, and so I have it to refer back to
later.

Change-Id: I5705dc3a4fc1f29db9031b6a4c3a2440a1dde513