cmd/compile: mark type of pointer to noalg type noalg too
Arrays marked noalg are created by the compiler to hold keys and values
to initialize map literals. The ssa backend creates a pointer type for the
array type when creating an OpAddr while processing the loop that
initializes the map from the arrays. The pointer type does not inherit
the noalg property but points to the noalg array type.

A similar problem occurred in #32595 for argument arrays of defer structs.

The fix here is more general and would apply to deferstructs marked noalg
too.

This fix does not prevent creating Slices, Structs, Chans, Arrays from
noalg types in the compiler. The broader and future-proof fix of disallowing
creation of any alged types depending on noalg types is left to after go1.17.

Fixes #47068

Change-Id: I99c365adbc0ae897d400efbd5c212dc9d185f7ff