sql: skip recording returnedAt for fixed conn pool
In github.com/tailscale/sqlite, this call to time.Now in a trivial
SELECT statement is nearly 8% of the total run time:

	name          old time/op  new time/op  delta
	EmptyExec-24   477ns ± 1%   441ns ± 1%  -7.47%  (p=0.000 n=20+20)

The value of returnedAt is only used to expire old connections, which
long-running servers using SQLite do not want to do. So do not record it
if we do not need it.

Change-Id: I770ec276c5631bcde079043667ec237dc8475aef