[dev.cmdgo] cmd/internal/str: add functions for quoting and splitting args
SplitQuotedFieldsAndUnescape takes a single string and splits it into
words on whitespace. Unlike strings.Fields, it respects single and
double quotes, allowing fields to contain space. It also allows
handles escaping with '\', unlike SplitQuotedFields, which is used in
cmd/go's -gcflags and a few others. SplitQuotedFieldsAndUnescape is a
copy of go/build.splitQuoted, which is used to parse #cgo directives.

JoinAndQuoteFields does the inverse: it accepts a list of strings that
may contain spaces, and it returns a string that can be parsed with
SplitQuotedFieldsAndUnescape.

QuotedStringListFlag uses these functions together to define new flags
that accept lists of arguments.

For golang/go#41400

Change-Id: I4986b753cb5e6fabb5b489bf26aedab889f853f5