net/http/httputil/reverseproxy: set the flush interval to -1 even if the Content-Type includes additional parameters
The existing implementation of flushInterval compared the entire
Content-Type header with "text/event-stream" to determine whether a
response needs to be flushed immediately. However, the Content-Type may
include an additional charset parameter as described here:

https://www.w3.org/TR/eventsource/#text-event-stream

With the additonal parameter, the header may look like
"text/event-stream;charset=utf-8". Thus, we parse the header and ignore
additional parameters before making the decision on whether or not the
response should be flushed.

Fixes #47359

Change-Id: Idccee54f55a64c67fe060820ca3880c51ff8bae2