[dev.typeparams] go/types: don't eagerly convert untyped RHS of shifts to uint
The spec does not explicitly state that we should convert an untyped RHS
of a shift expression to uint; it only says that it should be
representable by uint.

Fix this (as a proof of concept) by only checking representabilty while
checking the RHS, and relying on updateExprType to propagate the correct
type.

This results in new error messages and correct types for nested shift
expressions like 1.<<(2<<s).

DO NOT SUBMIT

Updates #47243

Change-Id: I655e1e283a983f6b32cf6abfd3823dd5439a8192