[dev.typeparams] cmd/compile: get rid of concretify use for bounds.
We just need to substitute from the typeparams to the shapes for the dst
type of the bound.

Removed concretify substituter, not used anymore. Also removed
shape2params, not needed anymore.

However, since the dst type is now not concrete, this gives more cases
where the linker can't find a method.

For a temporary solution, I realized we can change SymPtrWeak to SymPtr
in writeItab (see my change in writeItab). To get through all the tests,
I had to disable TestDeadcode.

I'm still working on the real solution. I realized that we need to call
MarkUsedIfaceMethod (or a new version of it) to mark a method as used on
a particular interface, else a type's method can be still deadcoded even
though MarkTypeUsedInInterface has been called on the concrete type.

Change-Id: Id67b72b350889dd3688b42739c337d5d79a0d1a2